backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",                # API Framework in Python
    "uvicorn[standard]>=0.34.0",       # Webserver for FastAPI
    "pydantic-settings>=2.2.1",   # For our configuration setup
    "asyncpg>=0.30.0",            # Async PostgreSQL client
    "argon2-cffi>=25.1.0",        # Password hashing
    "cryptography>=42.0.5",       # Cryptography for sessions
    "requests>=2.32.3",                # Requests, e.g. OIDC requests 
    "python-dotenv>=1.0.1",       # Allows for loading environment variables.
    "cassandra-driver",                # Python Cassandra driver
    "redis==6.2.0",                    # Redis client, allows us to connect to redis instance
    "aiosmtplib>=2.0.2",          # Async SMTP Client
    "email-validator>=2.1.1",     # Validates email fields

]

[dependency-groups]
dev = [
    "ruff>=0.11.8",                    # Code formatting
    "pytest",                          # Unit tests
    "httpx",                           # For API tests
]

[tool.ruff]
include = ["app/**/*.py"]
indent-width = 2

[tool.ruff.format]
quote-style="double"
indent-style="space"
line-ending="lf"